requirement.txt
numpy==1.22.4
numpy-stl==2.17.1
packaging==21.3
ply==3.11
PyOpenGL==3.1.6
pyparsing==3.0.9
PyQt6==6.3.0
PyQt6-Qt6==6.3.0
PyQt6-sip==13.3.1
pyqtgraph==0.12.4
python-utils==3.3.3
sip==6.6.1
toml==0.10.2
